requirements.txt
langchain
langchain-community
streamlit
python-dotenv
unstructured
langchain-groq
langchain-google-genai
langchain-huggingface
sentence_transformers
faiss-cpu